Struct logitech_led_sdk::Sdk

pub struct Sdk(_);

Entry to Api. This serves as proof of initalization.

Implementations

impl Sdk

pub fn new() -> Option<Self>

Create a new instance with no name. Returns None on failure.

pub fn new_with_name(name: &str) -> Option<Self>

Create a new instance with a name. Returns None on failure or if the passed name contains 0s.

pub fn get_version(&self) -> Option<(u32, u32, u32)>

Returns sdk version. Returns None on failure.

pub fn set_target(&self, target: Target) -> bool

Selects the target channels. Returns true if successful

pub fn set_lighting(&self, color: Color) -> bool

Sets the lighting. Returns true if successful.

pub fn set_lighting_for_key_with_name(&self, key: KeyName, color: Color) -> bool

Sets the lighting for a keyboard key by name

pub fn set_lighting_for_key_with_scan_code(
&self,
code: u32,
color: Color
) -> bool

Sets the lighting for a keyboard key by scan code

pub fn set_lighting_for_key_with_hid_code(
&self,
code: u32,
color: Color
) -> bool

Sets the lighting for a keyboard key by HID code

pub fn set_lighting_for_target_zone(
&self,
device: DeviceType,
zone: usize,
color: Color
) -> bool

Sets the lighting for a specific device's target zone

pub fn flash_lighting(
&self,
color: Color,
duration: Duration,
interval: Duration
) -> bool

Returns None if the call fails or any of the time values are too large. Duration how long the flashes occur overall. The interval is the time between flashes.

pub fn flash_single_key(
&self,
key: KeyName,
color: Color,
duration: Duration,
interval: Duration
) -> bool

Returns None if the call fails or any of the time values are too large. Duration how long the flashes occur overall. The interval is the time between flashes.

pub fn stop_effects(&self) -> bool

Stops all current LED effects

pub fn stop_effects_on_key(&self, key: KeyName) -> bool

Stops all LED effects on one key

pub fn pulse_lighting(
&self,
color: Color,
duration: Duration,
interval: Duration
) -> bool

Returns None if the call fails or any of the time values are too large. Duration how long the pulses occur overall. The interval is the time between pulses.

pub fn pulse_single_key(
&self,
key: KeyName,
start_color: Color,
finish_color: Color,
duration: Duration,
is_infinite: bool
) -> bool

Returns None if the call fails or any of the time values are too large. Duration how long the pulses occur overall.

pub fn save_lighting_for_key(&self, key: KeyName) -> bool

Saves the current lighting config for the given key

pub fn restore_lighting_for_key(&self, key: KeyName) -> bool

Saves the current lighting config for the given key